pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rio-kutty"
version = "0.1.0"
description = "FastAPI activity generation service"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "google-genai",
    "google-cloud-firestore>=2.14.0",
    "langgraph>=0.0.10",
    "langchain-core>=0.1.10",
    "Pillow>=10.2.0",
    "requests>=2.31.0",
    "httpx>=0.27.0",
    "pydantic-settings",
    "google-cloud-pubsub>=2.0.0",
    "google-cloud-storage>=2.14.0",
    "huggingface_hub>=0.1.0",
    "google-cloud-texttospeech>=2.16.0",
    "deepeval>=1.4.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]
dev = [
    "rio-kutty[test]",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
